first pass at adding paper authors as model authors
