+++ garden/src/features/gardens/components/garden-page/index.ts
@@ -5,6 +5,3 @@ export { default as ReviewNotice } from "./ReviewNotice";
 export { default as CitationBlock } from "./CitationBlock";
 export { default as GardenAssociatedMaterialsSection } from "./GardenAssociatedMaterialsSection";
 export { default as ShowAllMaterialsToggle } from "./ShowAllMaterialsToggle";
-export { default as AddMaterialWithFunctionSelect } from "./AddMaterialWithFunctionSelect"
-export { PaperCard, DatasetCard, RepositoryCard } from "./MaterialCards";
-export { NotebookCard } from "./components/NotebookCard";;
