+++ garden/src/features/materials/components/modals/MaterialModal.tsx
@@ -1,0 +1,10 @@
+import { Garden, ModalFunction } from "@/types";
+
+export interface MaterialModalProps {
+    context:  MaterialModalContext,
+}
+
+export interface MaterialModalContext {
+    garden?: Garden,
+    modalFunction?: ModalFunction,
+}
